// src/helpers/enums.ts
export enum ErrorMessage {
  NOT_FOUNT = 'Not found',
  USER_NOT_FOUNT = 'User not found',
  INVALID_INPUT = 'Invalid input',
  INVALID_USER_ID = 'Invalid user ID',
  METHOD_NOT_ALLOWED = 'Method not allowed',
  ENDROINT_NOT_FOUND = 'Endpoint not found',
  INTERNAL_SERVER_ERROR = 'Internal server error',
}